refactor flask app into modular blueprints with feature toggles
