+++ apps/legal_discovery/extensions.py
@@ -1,0 +1,4 @@
+from flask_socketio import SocketIO
+
+# Shared Socket.IO instance for the legal discovery app
+socketio = SocketIO()
